requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
Django==3.0
django-appconf==1.0.5
django-rq==2.5.1
django-video-encoding==1.0.0
future==0.18.2
idna==3.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
requests==2.27.1
rq==1.10.1
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.0
